by gacp

+++ Module-07/ex02/main.cpp
@@ -6,7 +6,7 @@
 /*   By: youmoukh <youmoukh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/05 16:42:01 by youmoukh          #+#    #+#             */
-/*   Updated: 2024/10/05 18:43:38 by youmoukh         ###   ########.fr       */
+/*   Updated: 2024/10/05 18:43:56 by youmoukh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ int main(int you, char**)
     std::cout << "Size of arr2: " << arr2.getSize() << std::endl;
 	
 	Array<std::string>str(3);
-	for (int i = 0x0;i <str.getSize();i++)
+	for (size_t i = 0x0; i <str.getSize() ; i++)
 	{
 		std::cout << i << std::endl;
 	}
